Еще правки по замечаниям
