+++ src/test/java/CreateUserTest.java
@@ -15,6 +15,7 @@ import io.qameta.allure.Step;
 
 import static constants.ApiConstants.BURGERS_URL;
 import static org.apache.http.HttpStatus.SC_FORBIDDEN;
+import static org.apache.http.HttpStatus.SC_OK;
 import static org.hamcrest.Matchers.equalTo;
 
 @RunWith(Parameterized.class)
@@ -24,7 +25,7 @@ public class CreateUserTest {
     private final String name;
     private User user;
     private UserRequests userRequests;
-    private String accessToken;
+    private Response createUserResponse;
 
     public CreateUserTest(String email, String password, String name) {
         this.email = email;
@@ -47,19 +48,18 @@ public class CreateUserTest {
         RestAssured.requestSpecification = RequestSpec.requestSpecification();
         user = new User(email, password, name);
         userRequests = new UserRequests();
-        accessToken = null; // Инициализация
+        createUserResponse = null;
     }
 
     @Test
     @Step("Создание пользователя")
     @DisplayName("Проверка создания пользователя")
     public void createUser() {
-        Response responseCreate = userRequests.createUser(user);
-        responseCreate.then()
-                .statusCode(200)
+        createUserResponse = userRequests.createUser(user);
+        createUserResponse.then()
+                .statusCode(SC_OK)
                 .assertThat()
                 .body("success", equalTo(true));
-        accessToken = responseCreate.then().extract().path("accessToken");
     }
 
     @Test
@@ -67,30 +67,33 @@ public class CreateUserTest {
     @Description("API должно возвращать ошибку 403 при попытке создать пользователя с уже занятым email")
     public void testCreateUserWithDuplicateEmail() {
         // Сначала создаем пользователя
-        Response firstResponse = userRequests.createUser(user);
-        accessToken = firstResponse.then().extract().path("accessToken");
+        createUserResponse = userRequests.createUser(user);
 
         // Пытаемся создать такого же пользователя еще раз
         Response duplicateResponse = userRequests.createUser(user);
 
         // Проверяем ответ сервера
         duplicateResponse.then()
-                .statusCode(SC_FORBIDDEN) // Ожидаем код 403
+                .statusCode(SC_FORBIDDEN)
                 .body("success", equalTo(false))
                 .body("message", equalTo("User already exists"));
     }
 
     @After
     public void deleteUser() {
-        if (accessToken != null) {
-            try {
-                Response responseDelete = userRequests.deleteUser(accessToken);
-                responseDelete.then()
-                        .statusCode(202)
-                        .body("success", equalTo(true));
-            } catch (Exception e) {
-                System.err.println("Ошибка при удалении пользователя: " + e.getMessage());
+        try {
+            // Получаем токен из ответа, если пользователь был создан
+            if (createUserResponse != null) {
+                String accessToken = createUserResponse.then().extract().path("accessToken");
+                if (accessToken != null) {
+                    Response responseDelete = userRequests.deleteUser(accessToken);
+                    responseDelete.then()
+                            .statusCode(202)
+                            .body("success", equalTo(true));
+                }
             }
+        } catch (Exception e) {
+            System.err.println("Ошибка при удалении пользователя: " + e.getMessage());
         }
     }
 }
